feature(header): add change server button for iOS
Jira-Id: VPNAPPL-1870

+++ apps/ios/ProtonVPN/Extensions/NSAttributedString+Extension.swift
@@ -40,7 +40,7 @@ extension NSAttributedString {
         return imageAttachment(image: image, size: size)
     }
     
-    static func imageAttachment(image: UIImage, baselineOffset: Int? = nil, size: CGSize? = nil) -> NSAttributedString {
+    static func imageAttachment(image: UIImage, baselineOffset: CGFloat? = nil, size: CGSize? = nil) -> NSAttributedString {
         let attachment = NSTextAttachment()
         if let size = size {
             attachment.bounds = CGRect(origin: .zero, size: size)
